# openapi.yaml
openapi: 3.1.0
info:
  title: ai-workflow-builder API
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    HTTP contract for the `@ai-workflow-builder/server` Express 4 service that
    powers workflow-builders.com.

    Turn a single natural-language prompt into a validated multi-agent workflow.
    Ambiguity is resolved by a deterministic **Grill-Me** spec loop *before* a
    workflow is scaffolded, and workflows are only persisted once they pass the
    aggregate's invariants (unique ids, resolvable dependencies, acyclic).

    This document was regenerated from the actual route table
    (`server/src/adapters/http/routes.js`) and the domain shapes it serves. It
    supersedes the earlier `/api/v1/agents|workflows` design, which described a
    deprecated Python/FastAPI product that no longer exists.

    ## Resource model
    The real aggregate root is a **Project** (`{ prompt, answers, spec }`). Each
    project has at most one compiled **Workflow**. There is no top-level
    `/agents` or `/workflows` collection.

    ## Lifecycle
    `POST /api/projects` (prompt) → `GET .../grill` → `POST .../answers`
    (repeat until `ready: true`) → `POST .../workflow/scaffold` →
    `PUT .../workflow` (edit + validate).

    ## Authentication & tenant isolation
    Every endpoint except `/api/health` requires a Clerk session JWT sent as
    `Authorization: Bearer <session token>`. The server verifies the token
...

    ## Key vault
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/projects:
    post:
      tags: [projects]
      summary: Create a project from a one-line prompt
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                  description: The one-line natural-language prompt.
                  example: summarise my emails
      responses:
        '201':
          description: Project created; an initial spec snapshot is derived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400':
          description: Prompt missing or not a non-empty string (`INVALID_PROMPT`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [projects]
      summary: List all projects
      operationId: listProjects
      responses:
        '200':
          description: Projects, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
  /api/projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      summary: Get a project
      operationId: getProject
      responses:
        '200':
          description: The project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [projects]
      summary: Delete a project (cascades to its workflow)
      operationId: deleteProject
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean, const: true }
                  id: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{id}/grill:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [grill]
      summary: Get the next Grill-Me questions and readiness
      operationId: grill
      parameters:
        - name: deep
          in: query
          required: false
          schema: { type: string, enum: ['true', 'false'] }
          description: >
            When `true`, also surface non-critical questions for dimensions that
            are already covered. Any value other than the literal string `true`
            is treated as false.
      responses:
        '200':
          description: Open questions plus per-dimension coverage.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GrillResult' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{id}/answers:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [grill]
      summary: Submit answers to grill questions
      operationId: answer
      description: >
        Merges the given answers into the project, re-derives the spec, and
        returns the updated project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: object
                  additionalProperties: { type: string }
                  description: Map of questionId → free-text answer.
                  example:
                    goal.outcome: a daily digest
                    inputs.source: my inbox
                    outputs.shape: markdown
                    success.measure: nothing urgent missed
      responses:
        '200':
          description: The updated project with a re-derived spec.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400':
          description: '`answers` is missing or not a plain object (`INVALID_ANSWERS`).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{id}/workflow/scaffold:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [workflow]
      summary: Compile the current spec into a starter workflow
      operationId: scaffoldWorkflow
      description: >
        Refuses with 409 unless the spec is ready, protecting against building
        on an under-specified spec. Pass `force: true` to override.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
                  description: Scaffold even if the spec is not ready.
      responses:
        '201':
          description: The scaffolded and persisted workflow.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workflow' }
        '409':
          description: Spec still has open questions (`SPEC_NOT_READY`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{id}/workflow:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [workflow]
      summary: Get the project's saved workflow
      operationId: getWorkflow
      responses:
        '200':
          description: The workflow, or `null` if none has been scaffolded/saved yet.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/Workflow' }
                  - { type: 'null' }
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [workflow]
      summary: Persist a user-edited workflow after validation
      operationId: saveWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workflow]
              properties:
                workflow: { $ref: '#/components/schemas/Workflow' }
      responses:
        '200':
          description: The persisted workflow.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workflow' }
        '422':
          description: Workflow failed invariant validation (`INVALID_WORKFLOW`). `details` lists every violation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/vault:
    get:
      tags: [vault]
      summary: List the org's stored provider keys (masked labels only)
      operationId: listVaultKeys
      description: >
        Returns masked labels and metadata for every vault entry in the
        caller's organization. Plaintext keys and wrapped key material are
        never returned.
      responses:
        '200':
          description: Masked vault entries, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/VaultEntry' }
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [vault]
      summary: Store a provider API key (envelope-encrypted)
      operationId: storeVaultKey
      description: >
        Encrypts the provider key with the org's Data Encryption Key (wrapped
        by the environment KEK) and stores it. The response contains only the
        masked preview.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, apiKey]
              properties:
                provider:
                  type: string
                  enum: [openai, anthropic, gemini, deepseek]
                  description: LLM provider.
                label:
                  type: string
                  description: Human-readable label (defaults to the provider name).
                apiKey:
                  type: string
                  description: The provider API key. Stored encrypted; never returned.
      responses:
        '201':
          description: Stored. Response carries masked metadata only.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaultEntry' }
        '400':
          description: 'Invalid provider or missing/oversized apiKey (`INVALID_PROVIDER` / `INVALID_API_KEY`).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/vault/{id}:
    parameters:
      - $ref: '#/components/parameters/VaultEntryId'
    get:
      tags: [vault]
      summary: Get one vault entry (masked)
      operationId: getVaultKey
      responses:
        '200':
          description: Masked vault entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaultEntry' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [vault]
      summary: Delete a vault entry (Owner only)
      operationId: deleteVaultKey
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean, const: true }
                  id: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # --- Increment 3: ecosystem catalog + Grill-Me SSE -----------------------
  # Catalog routes are read-only and require a session (any org — the catalog
  # is global MIT content). Writes are operator-only via the sync CLI, never
  # over HTTP.
...
  /api/catalog/personas:
    get:
      tags: [catalog]
      summary: Persona catalog grouped by division (Agent Marketplace)
      operationId: getPersonaCatalog
      responses:
        '200':
          description: Personas grouped by division with tool permission tags.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PersonaCatalog' }
        '404':
          description: The marketplace has not been synced yet (`CATALOG_EMPTY`).
...
  /api/catalog/agents:
    get:
      tags: [catalog]
      summary: List personas with division + tool permission tags
      operationId: listAgents
      parameters:
        - name: division
          in: query
          schema: { type: string }
          description: Filter by division id (e.g. `engineering`).
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive search over name, description, vibe.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: Persona records (latest good snapshot).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Persona' }
  /api/catalog/agents/{id}:
    get:
      tags: [catalog]
      summary: Fetch one persona (tool tags + markdown body)
      operationId: getAgent
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 'Full persona id, e.g. `agency-agents:engineering/backend-architect`.'
      responses:
        '200':
          description: The persona record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Persona' }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/catalog/lenses/{id}:
    get:
      tags: [catalog]
      summary: Fetch one lens (body + FIDELITY scorecard)
      operationId: getLens
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 'Full lens id, e.g. `nuwa-skill:munger-perspective`.'
      responses:
        '200':
          description: The lens record (fidelity present when a scorecard exists).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lens' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/catalog/snapshots:
    get:
      tags: [catalog]
      summary: Catalog sync provenance (snapshot history)
      operationId: listSnapshots
      parameters:
        - name: source
          in: query
          required: true
          schema: { type: string, enum: [agency-agents, nuwa-skill] }
      responses:
        '200':
          description: Snapshot rows, newest first (ok + failed attempts).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CatalogSnapshot' }
  /api/catalog/{source}:
    parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
          enum: [agency-agents, nuwa-skill]
    get:
      tags: [catalog]
      summary: Source alias — persona marketplace or lens catalog
      description: |
        `agency-agents` returns the persona catalog grouped by division;
        `nuwa-skill` returns the cognitive perspective lenses. Always serves
        the last-known-good snapshot.
      operationId: getCatalog
      responses:
        '200':
          description: Catalog payload.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/PersonaCatalog' }
                  - { $ref: '#/components/schemas/LensCatalog' }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/grill/stream:
    post:
      tags: [grill]
      summary: Open (or resume) the realtime Grill-Me SSE loop
      description: |
        Authenticated (Architect+). Opens a Server-Sent Events stream that
        interrogates ONE question at a time: `session` (identity + guardrail
        counters), `progress` (coverage), `question`, then `ready` +
        `compiled` (Python workflow) + `done` when the spec is covered.
        Answers are posted to `POST /api/grill/stream/{sessionId}/answers`.
        Turn (≤ 5) and token (≤ 15,000) ceilings are enforced in the service
        layer and surfaced as `limit` events / HTTP 429 — they cannot be
        bypassed by calling the API directly.
      operationId: openGrillStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: The workflow brief to interrogate (new session).
                projectId:
                  type: string
                  format: uuid
                  description: Attach the session to an existing project.
                sessionId:
                  type: string
                  description: Resume an existing session (replays its transcript).
      responses:
        '200':
          description: 'SSE event stream (`text/event-stream`).'
          content:
            text/event-stream:
              schema:
                type: string
                description: '`event: session|progress|question|ready|compiled|limit|done` frames, JSON `data` payloads.'
        '400':
          description: Missing prompt (`INVALID_PROMPT`).
        '404':
          description: Unknown session id on resume (`NOT_FOUND`).
        '429':
          description: Too many active sessions for the workspace (`TOO_MANY_SESSIONS`).
  /api/grill/stream/{sessionId}/answers:
    post:
      tags: [grill]
      summary: Answer the current grill question
      operationId: answerGrillStream
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answerId: { type: string, description: 'The question id from the last `question` event.' }
                text: { type: string }
              required: [answerId, text]
      responses:
        '200':
          description: Session state (counters + next position).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GrillStreamState' }
        '400':
          description: Malformed answer (`INVALID_ANSWER`).
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Session closed (`SESSION_CLOSED`).
        '429':
          description: Ceiling reached (`GRILL_LIMIT`) — turn or token budget exhausted.
  /api/grill/stream/{sessionId}:
    get:
      tags: [grill]
      summary: Read a grill stream session's state (reconnect probe)
      operationId: getGrillStreamState
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Session state JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GrillStreamState' }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token (verified against Clerk's JWKS).
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Project id (UUID).
...
  responses:
    NotFound:
      description: The referenced resource does not exist in this org (`NOT_FOUND`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      description: Uniform error envelope emitted by the HTTP adapter.
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          examples: [INVALID_PROMPT, INVALID_ANSWERS, NOT_FOUND, SPEC_NOT_READY, INVALID_WORKFLOW, INTERNAL]
        message: { type: string }
        details:
          description: Optional structured detail (e.g. the list of workflow validation errors).
      required: [error, message]
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        prompt: { type: string }
        answers:
          type: object
          additionalProperties: { type: string }
          description: Map of questionId → answer gathered so far.
        spec:
          oneOf:
            - { $ref: '#/components/schemas/Spec' }
            - { type: 'null' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, prompt, answers, createdAt, updatedAt]
...
    Workflow:
      type: object
      properties:
        id: { type: string, example: wf_9f1c… }
        name: { type: string }
        nodes:
          type: array
          items: { $ref: '#/components/schemas/WorkflowNode' }
      required: [nodes]
...